Update EnableMailboxAuditing.ps1
setting the parameters of mailbox auditing will actually reduce the log levels of the owner based on your command. So I updated it as when setting those parameters it will rewrite it and some level will be lost. Also in large environments running -resultsize unlimited and having it display all results won't work well +50,000 customers